refactor board interface and store
